+++ docs/CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog][kac] and this project adheres to
 quotes around code blocks in error output (#506)
 * an example/regression test for running background tasks without blocking the
   test run (#525, #535)
+* `bats_load_library` for loading libraries from the search path
+  `$BATS_LIB_PATH` (#548)
 
 ### Fixed